% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{File Size Options}
\alias{File Size Options}
\title{File Size Options}
\value{
Not applicable.
}
\description{
otel and otelsdk accept file size options in the following format:
\itemize{
\item As a positive numeric scalar, interpreted as number of bytes.
\item A string scalar with a positive number and a unit suffix.
Possible units: B, KB, KiB, MB, MiB, GB, GiB, TB, TiB, PB, PiB.
Units are case insensitive.
}
}
