% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentelemetry-package.R
\docType{package}
\name{opentelemetry-package}
\alias{opentelemetry}
\alias{opentelemetry-package}
\title{opentelemetry: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
