% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otelsdk-package.R
\docType{package}
\name{otelsdk-package}
\alias{otelsdk}
\alias{otelsdk-package}
\title{otelsdk: R SDK and Exporters for OpenTelemetry}
\value{
Not applicable.
}
\description{
OpenTelemetry is a collection of tools, APIs, and SDKs used to instrument, generate, collect, and export telemetry data (metrics, logs, and traces) for analysis in order to understand your software's performance and behavior. This package contains the OpenTelemetry SDK, and exporters. Use this package to export traces, metrics, logs from instrumented R code. Use the otel package to instrument your R code for OpenTelemetry.
}
\examples{
# Run your R script with OpenTelemetry tracing:
# OTEL_TRACER_EXPORTER=otlp R -f myapp.R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://otelsdk.r-lib.org}
  \item \url{https://github.com/r-lib/otelsdk}
  \item Report bugs at \url{https://github.com/r-lib/otelsdk/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Gábor Csárdi \email{csardi.gabor@gmail.com}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item  opentelemetry-cpp authors [contributor]
}

}
\keyword{internal}
