% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_default_tracer_provider}
\alias{set_default_tracer_provider}
\title{Set the default tracer provider}
\usage{
set_default_tracer_provider(tracer_provider)
}
\arguments{
\item{tracer_provider}{An OpenTelemetry tracer provider
(\code{opentelemetry_tracer_provider} object) to set as default.}
}
\value{
The previously set default tracer provider, or \code{NULL}, if no
default was set before.
}
\description{
Set the default tracer provider
}
