% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{setup_default_tracer}
\alias{setup_default_tracer}
\title{Get a tracer from the default tracer provider}
\usage{
setup_default_tracer(name)
}
\arguments{
\item{name}{Name of the new tracer. This is typically the R package
name.}
}
\value{
An OpenTelemetry tracer, an \code{opentelemetry_tracer} object.
}
\description{
Calls \code{\link[=get_default_tracer_provider]{get_default_tracer_provider()}} to get the default tracer
provider. Then calls its \verb{$get_tracer()} method to create a new tracer.
}
