% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{start_shiny_app}
\alias{start_shiny_app}
\title{Start tracing a Shiny app}
\usage{
start_shiny_app(service_name)
}
\arguments{
\item{service_name}{The name of the app.}
}
\value{
The OpenTelemetry tracer (\code{opentelemetry_tracer}), invisibly.
}
\description{
Call this function from \code{global.R}.
}
