% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{Time Interval Options}
\alias{Time Interval Options}
\title{Time Interval Options}
\value{
Not applicable.
}
\description{
otel and otelsdk accept time interval options in the following format:
\itemize{
\item A \link[base:difftime]{base::difftime} object.
\item A positive numeric scalar, interpreted as number of milliseconds.
It may be fractional.
\item A string scalar with a positive number and a time unit suffix.
Possible time units: us (microseconds), ms (milliseconds), s (seconds),
m (minutes), h (hours), d (days).
}

When the time interval is specified in an environment variable, it may be:
\itemize{
\item A positive number, interpreted as number of milliseconds.
It may be fractional.
\item A positive number with a time unit suffix.
Possible time units: us (microseconds), ms (milliseconds), s (seconds),
m (minutes), h (hours), d (days).
}
}
