% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-noop.R
\docType{data}
\name{tracer_provider_noop}
\alias{tracer_provider_noop}
\title{Tracer provider to ignore all tracing commands}
\format{
An object of class \code{list} of length 1.
}
\usage{
tracer_provider_noop
}
\description{
Tracer provider to ignore all tracing commands
}
\keyword{datasets}
