% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-stdout.R
\docType{data}
\name{tracer_provider_stdout}
\alias{tracer_provider_stdout}
\title{Tracer provider to write to the standard output}
\format{
An object of class \code{list} of length 1.
}
\usage{
tracer_provider_stdout
}
\description{
Tracer provider to write to the standard output
}
\keyword{datasets}
