% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{Environment Variables}
\alias{Environment Variables}
\title{Environment variables to configure otelsdk}
\value{
Not applicable.
}
\description{
See also the \link[otel:environmentvariables]{Environment Variables} in
the otel package, which is charge of selecting the exporters to use.
}
\details{
TODO
}
\examples{
# To start an R session using the OTLP exporter:
# OTEL_TRACES_EXPORTER=http R -q -f script.R
}
\seealso{
\link[otel:environmentvariables]{Environment Variables} in otel
}
