% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_default_tracer_provider}
\alias{get_default_tracer_provider}
\title{Get the default tracer provider}
\usage{
get_default_tracer_provider()
}
\value{
The default tracer provider, an \code{opentelemetry_tracer_provider}
object.
}
\description{
If there is no default set currently, then it creates and sets a
default.
}
\details{
The default tracer provider is created based on the
R_OPENTELEMETRY_DEFAULT_TRACER_PROVIDER environment variable.
The following values are allowed:
\itemize{
\item \code{stdout}: uses \link{tracer_provider_stdout}, to write traces to the
standard output.
\item \code{http}: uses \link{tracer_provider_http}, to send traces over HTTP.
\item \verb{<package>::<provider>}: will select the \verb{<provider>} object from
the \verb{<package>} package to use as a tracer provider. It calls
\verb{<package>::<provider>$new()} to create the new traver provider.
If this fails for some reason, e.g. the package is not installed,
then it throws an error.
}
}
