% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger-provider-file.R
\docType{data}
\name{logger_provider_file}
\alias{logger_provider_file}
\title{Logger provider to write log messages into a JSONL file.}
\format{
An object of class \code{list} of length 2.
}
\usage{
logger_provider_file
}
\description{
Logger provider to write log messages into a JSONL file.
}
\keyword{datasets}
