% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger-provider-http.R
\docType{data}
\name{logger_provider_http}
\alias{logger_provider_http}
\title{Logger provider to log over HTTP}
\format{
An object of class \code{list} of length 1.
}
\usage{
logger_provider_http
}
\description{
Logger provider to log over HTTP
}
\keyword{datasets}
