% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger-provider-stdout.R
\docType{data}
\name{logger_provider_stdstream}
\alias{logger_provider_stdstream}
\title{Logger provider to write to the standard output or standard error or
to a file}
\format{
An object of class \code{list} of length 2.
}
\usage{
logger_provider_stdstream
}
\description{
Logger provider to write to the standard output or standard error or
to a file
}
\keyword{datasets}
