% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter-provider-file.R
\docType{data}
\name{meter_provider_file}
\alias{meter_provider_file}
\title{Meter provider to collect metrics in JSONL files}
\format{
An object of class \code{list} of length 2.
}
\usage{
meter_provider_file
}
\description{
Meter provider to collect metrics in JSONL files
}
\keyword{datasets}
