% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter-provider-http.R
\docType{data}
\name{meter_provider_http}
\alias{meter_provider_http}
\title{Meter provider to send collected metrics over HTTP}
\format{
An object of class \code{list} of length 1.
}
\usage{
meter_provider_http
}
\description{
Meter provider to send collected metrics over HTTP
}
\keyword{datasets}
