% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter-provider-memory.R
\docType{data}
\name{meter_provider_memory}
\alias{meter_provider_memory}
\title{In-memory meter provider for debugging}
\format{
An object of class \code{list} of length 2.
}
\usage{
meter_provider_memory
}
\description{
In-memory meter provider for debugging
}
\keyword{datasets}
