% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter-provider-stdout.R
\docType{data}
\name{meter_provider_stdstream}
\alias{meter_provider_stdstream}
\title{Meter provider to write to the standard output or standard error or
to a file}
\format{
An object of class \code{list} of length 2.
}
\usage{
meter_provider_stdstream
}
\description{
Meter provider to write to the standard output or standard error or
to a file
}
\keyword{datasets}
