% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otelsdk-package.R
\docType{package}
\name{otelsdk-package}
\alias{otelsdk}
\alias{otelsdk-package}
\title{otelsdk: R Implementation of 'OpenTelemetry', Exporters}
\description{
High-quality, ubiquitous, and portable telemetry to enable effective observability. 'OpenTelemetry' is a collection of tools, 'APIs', and 'SDKs' used to instrument, generate, collect, and export telemetry data (metrics, logs, and traces) for analysis in order to understand your software's performance and behavior. Use this package to export traces, metrics, logs through the 'OpenTelemetry Protocol' ('OTLP').
}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item  opentelemetry-cpp authors [contributor]
}

}
\keyword{internal}
