% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-file.R
\docType{data}
\name{tracer_provider_file}
\alias{tracer_provider_file}
\title{Tracer provider to write traces into a JSONL file}
\format{
An object of class \code{list} of length 1.
}
\usage{
tracer_provider_file
}
\description{
Tracer provider to write traces into a JSONL file
}
\keyword{datasets}
