% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-http.R
\docType{data}
\name{tracer_provider_http}
\alias{tracer_provider_http}
\title{Tracer provider to export over HTTP}
\format{
An object of class \code{list} of length 2.
}
\usage{
tracer_provider_http
}
\description{
Tracer provider to export over HTTP
}
\keyword{datasets}
