% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-memory.R
\docType{data}
\name{tracer_provider_memory}
\alias{tracer_provider_memory}
\title{In-memory tracer provider for debugging}
\format{
An object of class \code{list} of length 2.
}
\usage{
tracer_provider_memory
}
\description{
In-memory tracer provider for debugging
}
\keyword{datasets}
