% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer-provider-stdout.R
\docType{data}
\name{tracer_provider_stdstream}
\alias{tracer_provider_stdstream}
\title{Tracer provider to write to the standard output or standard error or
to a file}
\format{
An object of class \code{list} of length 1.
}
\usage{
tracer_provider_stdstream
}
\description{
Tracer provider to write to the standard output or standard error or
to a file
}
\keyword{datasets}
